+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Significantly expand the built-in complex database by adding complexes from the Complex Portal and hu.MAP 2.0
 * Hide box plot legends if there are too many groups
+* Add MinProbGlobal imputation method
 
 # einprot 0.9.2